Refactored types into functional grouping folders instead of type (got rid of Interfaces), clean up on DeviceGestureService
