+++ Source/Prism.Tests/Mvvm/CompositeCommandFixture.cs
@@ -551,7 +551,7 @@ namespace Prism.Tests.Mvvm
             return true;
         }
 
-        public event EventHandler CanExecuteChanged;
+        public event EventHandler CanExecuteChanged = delegate { };
 
         public void Execute(object parameter)
         {
