+++ Source/PrismLibrary_Win10.sln
@@ -30,6 +30,7 @@ Global
 		{E6C50355-D01E-4CAA-884D-D7929861315C}.Debug|ARM.ActiveCfg = Debug|Any CPU
 		{E6C50355-D01E-4CAA-884D-D7929861315C}.Debug|ARM.Build.0 = Debug|Any CPU
 		{E6C50355-D01E-4CAA-884D-D7929861315C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E6C50355-D01E-4CAA-884D-D7929861315C}.Debug|x64.Build.0 = Debug|Any CPU
 		{E6C50355-D01E-4CAA-884D-D7929861315C}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{E6C50355-D01E-4CAA-884D-D7929861315C}.Debug|x86.Build.0 = Debug|Any CPU
 		{E6C50355-D01E-4CAA-884D-D7929861315C}.Release|ARM.ActiveCfg = Release|Any CPU
