--- Source/Windows10/Prism.Windows/AppModel/CancelableEventArgs.cs
@@ -1,20 +1,0 @@
-﻿using System;
-
-namespace Prism.Windows.AppModel
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public class CancelableEventArgs : EventArgs
-    {
-        /// <summary>
-        /// 
-        /// </summary>
-        public bool Cancel { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public CancelableEventArgs() { }
-    }
-}
